+++ example/spec/parallel/kinesisTests/KinesisTestErrorSpec.js
@@ -15,7 +15,8 @@ const {
   cleanupProviders,
   addCollections,
   cleanupCollections,
-  rulesList
+  readJsonFilesFromDir,
+  setProcessEnvironment
 } = require('@cumulus/integration-tests');
 
 const {
@@ -61,8 +62,10 @@ describe('The messageConsumer receives a bad record.\n', () => {
       console.log('Delete the Record from the queue.');
       await deleteSQSMessage(failureSqsUrl, this.ReceiptHandle);
     }
+
+    setProcessEnvironment(testConfig.stackName, testConfig.bucket);
     console.log(`\nDeleting ${ruleOverride.name}`);
-    const rules = await rulesList(testConfig.stackName, testConfig.bucket, ruleDirectory);
+    const rules = await readJsonFilesFromDir(ruleDirectory);
     // clean up stack state added by test
     console.log(`\nDeleting testStream '${streamName}'`);
     await Promise.all([