+++ packages/integration-tests/api/providers.js
@@ -24,7 +24,25 @@ const deleteProvider = (prefix, providerId) =>
     }
   });
 
+/**
+ * Fetch a provider from the Cumulus API
+ *
+ * @param {string} prefix - the prefix configured for the stack
+ * @param {string} providerId - a provider id
+ * @returns {Promise<Object>} - the provider fetched by the API
+ */
+const getProvider = (prefix, providerId) =>
+  callCumulusApi({
+    prefix,
+    payload: {
+      httpMethod: 'GET',
+      resource: '/{proxy+}',
+      path: `/providers/${providerId}`
+    }
+  }).then(({ body }) => JSON.parse(body));
+
 module.exports = {
   createProvider,
-  deleteProvider
+  deleteProvider,
+  getProvider
 };